Enum fann_sys::fann_errno_enum

pub enum fann_errno_enum {
    FANN_E_NO_ERROR,
    FANN_E_CANT_OPEN_CONFIG_R,
    FANN_E_CANT_OPEN_CONFIG_W,
    FANN_E_WRONG_CONFIG_VERSION,
    FANN_E_CANT_READ_CONFIG,
    FANN_E_CANT_READ_NEURON,
    FANN_E_CANT_READ_CONNECTIONS,
    FANN_E_WRONG_NUM_CONNECTIONS,
    FANN_E_CANT_OPEN_TD_W,
    FANN_E_CANT_OPEN_TD_R,
    FANN_E_CANT_READ_TD,
    FANN_E_CANT_ALLOCATE_MEM,
    FANN_E_CANT_TRAIN_ACTIVATION,
    FANN_E_CANT_USE_ACTIVATION,
    FANN_E_TRAIN_DATA_MISMATCH,
    FANN_E_CANT_USE_TRAIN_ALG,
    FANN_E_TRAIN_DATA_SUBSET,
    FANN_E_INDEX_OUT_OF_BOUND,
    FANN_E_SCALE_NOT_PRESENT,
}

Error events on fann and fann_train_data.

Variants

FANN_E_NO_ERROR

No error

FANN_E_CANT_OPEN_CONFIG_R

Unable to open configuration file for reading

FANN_E_CANT_OPEN_CONFIG_W

Unable to open configuration file for writing

FANN_E_WRONG_CONFIG_VERSION

Wrong version of configuration file

FANN_E_CANT_READ_CONFIG

Error reading info from configuration file

FANN_E_CANT_READ_NEURON

Error reading neuron info from configuration file

FANN_E_CANT_READ_CONNECTIONS

Error reading connections from configuration file

FANN_E_WRONG_NUM_CONNECTIONS

Number of connections not equal to the number expected

FANN_E_CANT_OPEN_TD_W

Unable to open train data file for writing

FANN_E_CANT_OPEN_TD_R

Unable to open train data file for reading

FANN_E_CANT_READ_TD

Error reading training data from file

FANN_E_CANT_ALLOCATE_MEM

Unable to allocate memory

FANN_E_CANT_TRAIN_ACTIVATION

Unable to train with the selected activation function

FANN_E_CANT_USE_ACTIVATION

Unable to use the selected activation function

FANN_E_TRAIN_DATA_MISMATCH

Irreconcilable differences between two fann_train_data structures

FANN_E_CANT_USE_TRAIN_ALG

Unable to use the selected training algorithm

FANN_E_TRAIN_DATA_SUBSET

Trying to take subset which is not within the training set

FANN_E_INDEX_OUT_OF_BOUND

Index is out of bound

FANN_E_SCALE_NOT_PRESENT

Scaling parameters not present