Enum fann_sys::fann_stopfunc_enum

pub enum fann_stopfunc_enum {
    FANN_STOPFUNC_MSE,
    FANN_STOPFUNC_BIT,
}

Stop criteria used during training.

Variants

FANN_STOPFUNC_MSE

Stop criterion is Mean Square Error (MSE) value.

FANN_STOPFUNC_BIT

Stop criterion is number of bits that fail. The number of bits means the number of output neurons which differ more than the bit fail limit (see fann_get_bit_fail_limit, fann_set_bit_fail_limit). The bits are counted in all of the training data, so this number can be higher than the number of training data.