fann_sys::fann_create_sparse

pub unsafe extern fn fann_create_sparse(connection_rate: c_float, num_layers: c_uint) -> *mut fann

Creates a standard backpropagation neural network, which is not fully connected.

Parameters

connection_rate - The connection rate controls how many connections there will be in the network. If the connection rate is set to 1, the network will be fully connected, but if it is set to 0.5, only half of the connections will be set. A connection rate of 1 will yield the same result as fann_create_standard.

num_layers - The total number of layers including the input and the output layer.

... - Integer values determining the number of neurons in each layer starting with the input layer and ending with the output layer.

Returns

A pointer to the newly created fann.

See also

fann_create_sparse_array, fann_create_standard, fann_create_shortcut

This function appears in FANN >= 2.0.0.