fann_sys::fann_get_activation_function

pub unsafe extern fn fann_get_activation_function(ann: *const fann, layer: c_int, neuron: c_int) -> fann_activationfunc_enum

Get the activation function for neuron number neuron in layer number layer, counting the input layer as layer 0.

It is not possible to get activation functions for the neurons in the input layer.

Information about the individual activation functions is available at fann_activationfunc_enum.

Returns

The activation function for the neuron or FANN_NONE if the neuron is not defined in the neural network.

See also

fann_set_activation_function_layer, fann_set_activation_function_hidden, fann_set_activation_function_output, fann_set_activation_steepness, fann_set_activation_function

This function appears in FANN >= 2.1.0.