fann_sys::fann_get_activation_steepness

pub unsafe extern fn fann_get_activation_steepness(ann: *const fann, layer: c_int, neuron: c_int) -> fann_type

Get the activation steepness for neuron number neuron in layer number layer, counting the input layer as layer 0.

It is not possible to get activation steepness for the neurons in the input layer.

The steepness of an activation function says something about how fast the activation function goes from the minimum to the maximum. A high value for the activation function will also give a more aggressive training.

When training neural networks where the output values should be at the extremes (usually 0 and 1, depending on the activation function), a steep activation function can be used (e.g. 1.0).

The default activation steepness is 0.5.

Returns

The activation steepness for the neuron or -1 if the neuron is not defined in the neural network.

See also

fann_set_activation_steepness_layer, fann_set_activation_steepness_hidden, fann_set_activation_steepness_output, fann_set_activation_function, fann_set_activation_steepness

This function appears in FANN >= 2.1.0