fann_sys::fann_get_cascade_candidate_change_fraction

pub unsafe extern fn fann_get_cascade_candidate_change_fraction(ann: *const fann) -> c_float

The cascade candidate change fraction is a number between 0 and 1 determining how large a fraction the fann_get_MSE value should change within fann_get_cascade_candidate_stagnation_epochs during training of the candidate neurons, in order for the training not to stagnate. If the training stagnates, the training of the candidate neurons will be ended and the best candidate will be selected.

This means: If the MSE does not change by a fraction of fann_get_cascade_candidate_change_fraction during a period of fann_get_cascade_candidate_stagnation_epochs, the training of the candidate neurons is stopped because the training has stagnated.

If the cascade candidate change fraction is low, the candidate neurons will be trained more and if the fraction is high they will be trained less.

The default cascade candidate change fraction is 0.01, which is equivalent to a 1% change in MSE.

See also

fann_set_cascade_candidate_change_fraction, fann_get_MSE, fann_get_cascade_candidate_stagnation_epochs

This function appears in FANN >= 2.0.0.