fann_sys::fann_save_to_fixed

pub unsafe extern fn fann_save_to_fixed(ann: *mut fann, configuration_file: *const c_char) -> c_int

Saves the entire network to a configuration file. But it is saved in fixed point format no matter which format it is currently in.

This is useful for training a network in floating points, and then later executing it in fixed point.

The function returns the bit position of the fix point, which can be used to find out how accurate the fixed point network will be. A high value indicates high precision, and a low value indicates low precision.

A negative value indicates very low precision, and a very strong possibility for overflow. (the actual fix point will be set to 0, since a negative fix point does not make sense).

Generally, a fix point lower than 6 is bad, and should be avoided. The best way to avoid this is to have fewer connections to each neuron, or just fewer neurons in each layer.

The fixed point use of this network is only intended for use on machines that have no floating point processor, like an iPAQ. On normal computers the floating point version is actually faster.

See also

fann_create_from_file, fann_save

This function appears in FANN >= 1.0.0.