fann_sys::fann_save_train_to_fixed

pub unsafe extern fn fann_save_train_to_fixed(data: *mut fann_train_data, filename: *const c_char, decimal_point: c_uint) -> c_int

Saves the training structure to a fixed point data file.

This function is very useful for testing the quality of a fixed point network.

Return

The function returns 0 on success and -1 on failure.

See also

fann_save_train

This function appears in FANN >= 1.0.0.