fann_sys::fann_set_activation_steepness

pub unsafe extern fn fann_set_activation_steepness(ann: *mut fann, steepness: fann_type, layer: c_int, neuron: c_int)

Set the activation steepness for neuron number neuron in layer number layer, counting the input layer as layer 0.

It is not possible to set activation steepness for the neurons in the input layer.

The steepness of an activation function says something about how fast the activation function goes from the minimum to the maximum. A high value for the activation function will also give a more aggressive training.

When training neural networks where the output values should be at the extremes (usually 0 and 1, depending on the activation function), a steep activation function can be used (e.g. 1.0).

The default activation steepness is 0.5.

See also

fann_set_activation_steepness_layer, fann_set_activation_steepness_hidden, fann_set_activation_steepness_output, fann_set_activation_function, fann_get_activation_steepness

This function appears in FANN >= 2.0.0.