fann_sys::fann_set_error_log

pub unsafe extern fn fann_set_error_log(errdat: *mut fann_error, log_file: *mut FILE)

Change where errors are logged to. Both fann and fann_data can be cast to fann_error, so this function can be used to set either of these.

If log_file is NULL, no errors will be printed.

If errdat is NULL, the default log will be set. The default log is the log used when creating fann and fann_data. This default log will also be the default for all new structs that are created.

The default behavior is to log them to stderr.

See also

fann_error

This function appears in FANN >= 1.1.0.