fann_sys::fann_set_output_scaling_params

pub unsafe extern fn fann_set_output_scaling_params(ann: *mut fann, data: *const fann_train_data, new_output_min: c_float, new_output_max: c_float) -> c_int

Calculate output scaling parameters for future use based on training data.

Parameters

ann - ANN for which parameters need to be calculated

data - training data that will be used to calculate scaling parameters

new_output_min - desired lower bound in output data after scaling (not strictly followed)

new_output_max - desired upper bound in output data after scaling (not strictly followed)

See also

fann_set_input_scaling_params

This function appears in FANN >= 2.1.0