fann_sys::fann_set_user_data

pub unsafe extern fn fann_set_user_data(ann: *mut fann, user_data: *mut c_void)

Store a pointer to user defined data. The pointer can be retrieved with fann_get_user_data for example in a callback. It is the user's responsibility to allocate and deallocate any data that the pointer might point to.

Parameters

ann - A previously created neural network structure of type fann pointer.

user_data - A void pointer to user defined data.

This function appears in FANN >= 2.1.0.