fann_sys::fann_train_on_data

pub unsafe extern fn fann_train_on_data(ann: *mut fann, data: *const fann_train_data, max_epochs: c_uint, epochs_between_reports: c_uint, desired_error: c_float)

Trains on an entire dataset, for a period of time.

This training uses the training algorithm chosen by fann_set_training_algorithm, and the parameters set for these training algorithms.

Parameters

ann - The neural network

data - The data that should be used during training

max_epochs - The maximum number of epochs the training should continue

epochs_between_reports - The number of epochs between printing a status report to stdout. A value of zero means no reports should be printed.

desired_error - The desired fann_get_MSE or fann_get_bit_fail, depending on which stop function is chosen by fann_set_train_stop_function.

Instead of printing out reports every epochs_between_reports, a callback function can be called (see fann_set_callback).

See also

fann_train_on_file, fann_train_epoch

This function appears in FANN >= 1.0.0.