Struct fann_sys::fann

pub struct fann {
    // some fields omitted
}

The fast artificial neural network (fann) structure.

Data within this structure should never be accessed directly, but only by using the fann_get_... and fann_set_... functions.

The fann structure is created using one of the fann_create_... functions and each of the functions which operates on the structure takes a fann pointer as the first parameter.

See also

fann_create_standard, fann_destroy